# docs/pricing-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Farmacore — Pricing API",
    "version": "1.0.0",
    "description": "API REST do módulo de precificação (multi-tenant por JWT). Sem prefixo global; o tenant vem do token (não há `:tenant` na URL nem header de tenant). Todas as rotas exigem JWT salvo as marcadas como públicas. RBAC por `x-roles` (admin/operator/viewer): role insuficiente → 403. ValidationPipe global (whitelist + forbidNonWhitelisted) → 400 para body/query inválido ou campo extra. `:id` é UUID (ParseUUIDPipe → 400 se malformado)."
  },
  "servers": [
    {
      "url": "/",
      "description": "Mesma origem da API (paths exatos dos @Controller)"
    }
  ],
...
  "paths": {
...
    "/pricing/suggestion-rules": {
      "get": {
        "tags": [
          "suggestion-rules"
        ],
        "summary": "Lista regras",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "operator",
          "admin"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SuggestionRuleApi"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Sem JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Role insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "suggestion-rules"
        ],
        "summary": "Cria regra (auditado)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "operator",
          "admin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertSuggestionRuleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Criada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestionRuleApi"
                }
              }
            }
          },
          "400": {
            "description": "Validação (XOR classe/cluster, concorrente duplicado/inválido, peso, etc.)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Role insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/pricing/suggestion-rules/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "patch": {
        "tags": [
          "suggestion-rules"
        ],
        "summary": "Atualiza regra (auditado)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "operator",
          "admin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertSuggestionRuleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestionRuleApi"
                }
              }
            }
          },
          "400": {
            "description": "Validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Regra não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "suggestion-rules"
        ],
        "summary": "Remove regra (soft-delete, auditado)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "operator",
          "admin"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "deleted": {
                      "type": "boolean",
                      "const": true
                    }
                  },
                  "required": [
                    "id",
                    "deleted"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Regra não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Envelope de erro padrão do NestJS",
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "CompetitorOrigin": {
        "type": "string",
        "enum": [
          "DROGAL",
          "DROGASIL",
          "PAGUE_MENOS",
          "IKESAKI",
          "MICHELASSI",
          "PACHECO",
          "SAO_PAULO",
          "VENANCIO",
          "INDIANA"
        ]
      },
...
      "SuggestionStrategy": {
        "type": "string",
        "enum": [
          "margem",
          "concorrencia"
        ]
      },
      "CompetitorMode": {
        "type": "string",
        "enum": [
          "weighted",
          "cascade",
          "lowest"
        ]
      },
...
      "RuleCompetitor": {
        "type": "object",
        "required": [
          "competitor",
          "weight"
        ],
        "properties": {
          "competitor": {
            "$ref": "#/components/schemas/CompetitorOrigin"
          },
          "weight": {
            "type": "number",
            "description": "% em weighted; 1 em cascade/lowest"
          }
        }
      },
      "RuleCompetitorInput": {
        "type": "object",
        "required": [
          "competitor"
        ],
        "properties": {
          "competitor": {
            "$ref": "#/components/schemas/CompetitorOrigin"
          },
          "weight": {
            "type": "number",
            "minimum": 0,
            "maximum": 100,
            "description": "obrigatório/efetivo só em weighted"
          }
        }
      },
      "SuggestionRuleApi": {
        "type": "object",
        "required": [
          "id",
          "name",
          "classifications",
          "clusterId",
          "clusterName",
          "excludeClusterIds",
          "strategy",
          "minMargin",
          "competitorMode",
          "competitors",
          "variationPct",
          "noCompetitorMargin",
          "priceControlled",
          "ignorePbm",
          "blockPbmInMargin",
          "cascadeByPriority",
          "applyRounding",
          "active",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "classifications": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "clusterId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "clusterName": {
            "type": "string",
            "nullable": true
          },
          "excludeClusterIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "strategy": {
            "$ref": "#/components/schemas/SuggestionStrategy"
          },
          "minMargin": {
            "type": "number"
          },
          "competitorMode": {
            "$ref": "#/components/schemas/CompetitorMode"
          },
          "competitors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RuleCompetitor"
            }
          },
          "variationPct": {
            "type": "number"
          },
          "noCompetitorMargin": {
            "type": "number",
            "nullable": true
          },
          "priceControlled": {
            "type": "boolean"
          },
          "ignorePbm": {
            "type": "boolean"
          },
          "blockPbmInMargin": {
            "type": "boolean",
            "description": "Bloqueia PBM também na estratégia margem (default false)"
          },
          "cascadeByPriority": {
            "type": "boolean",
            "description": "Cascade segue a priority do tenant (default false)"
          },
          "applyRounding": {
            "type": "boolean"
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UpsertSuggestionRuleDto": {
        "type": "object",
        "required": [
          "name",
          "minMargin"
        ],
        "description": "Validações cruzadas no service: classificações XOR cluster; concorrência exige ≥1 concorrente; concorrente duplicado → 400.",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "classifications": {
            "type": "array",
            "maxItems": 200,
            "items": {
              "type": "string",
              "maxLength": 200
            }
          },
          "clusterId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "excludeClusterIds": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "strategy": {
            "$ref": "#/components/schemas/SuggestionStrategy"
          },
          "minMargin": {
            "type": "number",
            "minimum": 0,
            "maximum": 95
          },
          "competitorMode": {
            "$ref": "#/components/schemas/CompetitorMode"
          },
          "competitors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RuleCompetitorInput"
            }
          },
          "variationPct": {
            "type": "number",
            "minimum": -90,
            "maximum": 90
          },
          "noCompetitorMargin": {
            "type": "number",
            "minimum": 0,
            "maximum": 95,
            "nullable": true
          },
          "priceControlled": {
            "type": "boolean"
          },
          "ignorePbm": {
            "type": "boolean"
          },
          "blockPbmInMargin": {
            "type": "boolean"
          },
          "cascadeByPriority": {
            "type": "boolean"
          },
          "applyRounding": {
            "type": "boolean"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}